# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/spaces": {
      "get": {
        "operationId": "SpacesController_getAll",
        "summary": "Lihat Semua Space Coworking (Filter ?tipe & ?search)",
        "parameters": [
          {
            "name": "tipe",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "desk",
                "meeting_room",
                "private_office"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Spaces (Katalog & Ketersediaan)"
        ]
      }
    },
    "/api/spaces/{id}": {
      "get": {
        "operationId": "SpacesController_getById",
        "summary": "Lihat Detail Space Coworking Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Spaces (Katalog & Ketersediaan)"
        ]
      }
    },
...
    "/api/diskon/{id}": {
      "get": {
        "operationId": "DiskonController_getById",
        "summary": "Lihat Detail Diskon Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diskon & Promo"
        ]
      }
    },
    "/api/reservasi": {
      "post": {
        "operationId": "ReservasiController_create",
        "summary": "Buat Pemesanan Space Baru (Member)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reservasi Member"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/admin/profile": {
      "get": {
        "operationId": "AdminController_getProfile",
        "summary": "Lihat Data Profil Lokasi Coworking Space",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_updateProfile",
        "summary": "Update Data Profil Lokasi Coworking Space",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCoworkingProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/members": {
      "get": {
        "operationId": "AdminController_getMembers",
        "summary": "Daftar Semua Member / Pelanggan Coworking",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "AdminController_createMember",
        "summary": "Tambah Data Member Baru oleh Admin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/members/{id}": {
      "get": {
        "operationId": "AdminController_getMemberById",
        "summary": "Detail Data Member Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_updateMember",
        "summary": "Update Data Member / Pelanggan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteMember",
        "summary": "Hapus Data Member / Pelanggan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/spaces": {
      "get": {
        "operationId": "AdminController_getSpaces",
        "summary": "Daftar Semua Ruangan & Meja Milik Admin",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "AdminController_createSpace",
        "summary": "Tambah Ruangan / Meja Space Baru",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/spaces/{id}": {
      "get": {
        "operationId": "AdminController_getSpaceById",
        "summary": "Detail Data Space Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_updateSpace",
        "summary": "Update Data Ruangan & Fasilitas Space",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteSpace",
        "summary": "Hapus Data Ruangan / Meja Space",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/diskon": {
      "get": {
        "operationId": "AdminController_getDiskons",
        "summary": "Daftar Semua Kode Promo / Diskon Event",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "AdminController_createDiskon",
        "summary": "Tambah Kode Promo / Diskon Event Baru",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiskonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/diskon/{id}": {
      "get": {
        "operationId": "AdminController_getDiskonById",
        "summary": "Detail Data Diskon Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_updateDiskon",
        "summary": "Update Data Kode Promo & Periode Diskon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiskonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteDiskon",
        "summary": "Hapus Kode Promo / Diskon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/reservasi": {
      "get": {
        "operationId": "AdminController_getReservasis",
        "summary": "Lihat Seluruh Reservasi Coworking (?month, ?year, ?status, ?id_space, ?tanggal)",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "belum_dikonfirm",
                "disetujui",
                "aktif",
                "selesai",
                "dibatalkan"
              ],
              "type": "string"
            }
          },
          {
            "name": "id_space",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tanggal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/reservasi/{id}/status": {
      "patch": {
        "operationId": "AdminController_updateReservasiStatus",
        "summary": "Konfirmasi & Ubah Status Pemesanan (disetujui, dibatalkan, dll)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservasiStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/reservasi/{id}/check-in": {
      "post": {
        "operationId": "AdminController_checkIn",
        "summary": "Check-In Pelanggan (Status Berubah ke Aktif/Digunakan)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/reservasi/{id}/check-out": {
      "post": {
        "operationId": "AdminController_checkOut",
        "summary": "Check-Out Pelanggan (Status Berubah ke Selesai)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/reports/monthly": {
      "get": {
        "operationId": "AdminController_getMonthlyReport",
        "summary": "Rekapitulasi Estimasi & Realisasi Pendapatan Per Bulan",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/admin/reports/income": {
      "get": {
        "operationId": "AdminController_getIncome",
        "summary": "Alias Rekapitulasi Pendapatan Bulanan",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Space Booking API",
    "description": "Backend RESTful API untuk Sistem Reservasi Coworking Space - UKK RPL 2026/2027 SMK Telkom Malang",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateReservasiDto": {
        "type": "object",
        "properties": {
          "id_space": {
            "type": "number",
            "example": 1
          },
          "tanggal_reservasi": {
            "type": "string",
            "example": "2026-08-30"
          },
          "jam_mulai": {
            "type": "string",
            "example": "09:00"
          },
          "durasi_jam": {
            "type": "number",
            "example": 3
          },
          "id_diskon": {
            "type": "number",
            "example": 1
          },
          "kode_promo": {
            "type": "string",
            "example": "DISKONHEMAT20"
          }
        },
        "required": [
          "id_space",
          "tanggal_reservasi",
          "jam_mulai",
          "durasi_jam"
        ]
      },
      "UpdateCoworkingProfileDto": {
        "type": "object",
        "properties": {
          "nama_coworking": {
            "type": "string",
            "example": "Moklet Hub Coworking Space"
          },
          "nama_pemilik": {
            "type": "string",
            "example": "Ahmad Bidin, S.Kom"
          },
          "telp": {
            "type": "string",
            "example": "081298765432"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Telkom No. 1, Malang"
          },
          "deskripsi": {
            "type": "string",
            "example": "Coworking space terbaik di Malang"
          }
        },
        "required": [
          "nama_coworking",
          "nama_pemilik",
          "telp"
        ]
      },
      "CreateMemberAdminDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user_budi"
          },
          "password": {
            "type": "string",
            "example": "Secret123!"
          },
          "nama_member": {
            "type": "string",
            "example": "Budi Raharjo"
          },
          "instansi": {
            "type": "string",
            "example": "SMK Telkom Malang"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Danau Ranau No. 1, Malang"
          },
          "telp": {
            "type": "string",
            "example": "085712345678"
          },
          "foto": {
            "type": "string",
            "example": "budi_raharjo.jpg"
          }
        },
        "required": [
          "username",
          "password",
          "nama_member",
          "instansi",
          "alamat",
          "telp"
        ]
      },
      "UpdateMemberAdminDto": {
        "type": "object",
        "properties": {
          "nama_member": {
            "type": "string",
            "example": "Budi Raharjo, S.T."
          },
          "instansi": {
            "type": "string",
            "example": "PT Teknologi Hebat"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Danau Ranau No. 2, Malang"
          },
          "telp": {
            "type": "string",
            "example": "085712345678"
          },
          "password": {
            "type": "string",
            "example": "NewSecret123!"
          },
          "foto": {
            "type": "string",
            "example": "budi_new.jpg"
          }
        }
      },
      "CreateSpaceDto": {
        "type": "object",
        "properties": {
          "nama_space": {
            "type": "string",
            "example": "Personal Desk Alpha 01"
          },
          "harga_per_jam": {
            "type": "number",
            "example": 25000
          },
          "tipe": {
            "type": "string",
            "example": "desk",
            "enum": [
              "desk",
              "meeting_room",
              "private_office"
            ]
          },
          "kapasitas": {
            "type": "number",
            "example": 1
          },
          "deskripsi": {
            "type": "string",
            "example": "WiFi 100Mbps, stopkontak, coffee"
          },
          "foto": {
            "type": "string",
            "example": "desk_alpha_01.jpg"
          }
        },
        "required": [
          "nama_space",
          "harga_per_jam",
          "tipe",
          "kapasitas",
          "deskripsi"
        ]
      },
...
    }
  }
}